DictionarySearchResults in separate component
